change in event.html and story.html
I've made some modification in the event page the " date " which is mentioned right below the "
Learning mental wellness card content " was overlapped with the above content.

I've also made a change in the story page - the logo of the title "Moksha" was little bit disturbed so I tried to fix it.